Refactor code to adhere to clean architecture
Move code to separate folders for interfaces and the actual
implementations. Dependancies between the different structs are
minimized as the implementations only depend on other interfaces. The
only difference is the RotatingCleaningPlan struct which after marking a
Task as done creates a new Cleaningjob struct instead of using a generic
interface.
The pseudo REST implementation is replaced by an actual REST handler
using the gorilla/mux package.