Fix parsing from JSON file
This extends the JSON Converter struct to read from a saved file and
Unmarshal the structs correctly. Sadly there seems to be no other way
than telling the function which struct exactly it should use. This leads
further away from a 100% clean architecture. There are two occurences in
the Converter class where explicit types are used.
Also the types are a bit simpler and not boxed to make for easier
parsing.