Clean up JSON response
Up until now the webservice always returned the whole cleaning plan
including all the people in it if asked for tasks. If asked to filter
tasks by assignee, the service would return all task fields including
the assignee him-/herself everytime.
Now only the tasks are returned for a 'http.GET:/tasks' and the assignee
is removed before returning the filtered tasks for a
'http.GET:/tasks/<assigneename>'.